+++ src/main/kotlin/com/devbabys/shoppingmall/Controller/IndexController.kt
@@ -1,5 +1,6 @@
 package com.devbabys.shoppingmall.Controller
 
+import com.devbabys.shoppingmall.Cookie.CookieUtil
 import jakarta.servlet.http.HttpServletRequest
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Controller
@@ -9,13 +10,13 @@ import org.springframework.web.bind.annotation.GetMapping
 @Controller
 class IndexController(
     @Autowired
-    private val cookieController: CookieController
+    private val cookieUtil: CookieUtil
 ) {
 
     @GetMapping("/")
     fun index(model: Model, request: HttpServletRequest) : String {
-        var token = cookieController.getToken(request, "token")
-        println("########## IndexController : / : [token] $token ##########")
+        var token = cookieUtil.getToken(request, "token");
+        model.addAttribute("token", token)
         return "index"
     }
 }
