+++ src/main/kotlin/com/devbabys/shoppingmall/Security/JwtUtil.kt
@@ -10,11 +10,6 @@ import kotlin.collections.HashMap
 
 @Component
 class JwtUtil {
-//    @Value("\${jwt.secret.key}")
-//    private lateinit var secret: String
-//    @Value("\${jwt.expiration.date}")
-//    private lateinit var expiration: String
-
     private val secretKey = Jwts.SIG.HS512.key().build();
 
     fun generateToken(email: String): String {
@@ -49,15 +44,5 @@ class JwtUtil {
     fun validateToken(token: String, email: String): Boolean {
         val extractedEmail = extractedEmail(token)
         return (extractedEmail == email && !isTokenExpired(token))
-    }
-
-    fun getAfter30Days(time: Int): Date {
-        val now = Date()
-        val calendar = Calendar.getInstance()
-        calendar.time = now
-        calendar.add(Calendar.DAY_OF_MONTH, time) // 현재 시간에서 30일 후
-        val expirationDate = calendar.time
-
-        return expirationDate
     }
 }
