+++ src/main/resources/templates/index.html
@@ -11,9 +11,6 @@
     <!-- 인덱스 페이지의 내용 -->
     <center>
         <h2>Content of Index Page</h2>
-        <div th:if="${token != null}">
-            <span> 로그인 상태 </span>
-        </div>
     </center>
     <!-- 인덱스 페이지의 내용 추가 -->
